Update SPLTheory.hs
Refinement of the top-left quadrant, considering concrete implementation for partial composition of models and expressions, which are leveraged by pi' and sigma'. Again, some isomorphism seems to exist between compositional model and compositional expressions. Since sigma' and pi' have similar behavior, this seems to guarantee the commutativity. A conjecture is that the proof should have be same structure than of the top-right quadrant.